+++ src/01-one/lexicon/methods/index.js
@@ -1,7 +1,7 @@
 import { toPresent, toPast, toFuture, toImperative, toGerund, isPerfective, getAspectPair, all as allVerb } from './verbs/conjugate.js'
 import { fromPresent, fromPast, fromImperative } from './verbs/to-root.js'
-import { toPlural, toSingular } from './nouns/inflect.js'
-import { toMasculine, toFeminine, toNeuter, toPlural as toPluralAdj } from './adjectives/agree.js'
+import { toPlural, toSingular, decline, guessGender } from './nouns/inflect.js'
+import { toMasculine, toFeminine, toNeuter, toPlural as toPluralAdj, toComparative, toSuperlative } from './adjectives/agree.js'
 
 export default {
   verb: {
@@ -20,11 +20,15 @@ export default {
   noun: {
     toPlural,
     toSingular,
+    decline,
+    guessGender,
   },
   adjective: {
     toMasculine,
     toFeminine,
     toNeuter,
     toPlural: toPluralAdj,
+    toComparative,
+    toSuperlative,
   },
 }