+++ src/01-one/tagset/tags/nouns.js
@@ -103,11 +103,15 @@ export default {
 
   FemaleNoun: {
     is: 'Noun',
-    not: ['MaleNoun']
+    not: ['MaleNoun', 'NeuterNoun']
   },
   MaleNoun: {
     is: 'Noun',
-    not: ['FemaleNoun']
+    not: ['FemaleNoun', 'NeuterNoun']
+  },
+  NeuterNoun: {
+    is: 'Noun',
+    not: ['MaleNoun', 'FemaleNoun']
   },
 
 }